Add dark overlay shimmer effect when registering
