+++ apps/web/src/components/Basenames/UsernamePill/types.ts
@@ -1,0 +1,12 @@
+import { BaseName } from '@coinbase/onchainkit/identity';
+import { Address } from 'viem';
+
+export enum UsernamePillVariants {
+  Inline = 'inline',
+  Card = 'card',
+}
+export type UsernamePillProps = {
+  variant: UsernamePillVariants;
+  username: BaseName;
+  address?: Address;
+};
